Add views for user CRUD routes

+++ handlers/user.py
@@ -29,7 +29,7 @@ def change_password():
 
             if password1 != password2:
                 flash("Passwords does not match!", "danger")
-                return redirect(url_for('user.profile'))
+                return redirect(url_for('user.change_password'))
 
             logged_user.password = hashlib.sha256(password1.encode()).hexdigest()
             db.add(logged_user)
@@ -38,8 +38,10 @@ def change_password():
             flash("Password changed successfully", "success")
             return redirect(url_for('user.profile'))
 
+        if request.method == 'GET':
+            return render_template('change_password.html')
     else:
-        redirect(url_for('auth.login'))
+        return redirect(url_for('auth.login'))
 
 
 @user.route('/change-username', methods=["GET", "POST"])
@@ -56,20 +58,29 @@ def change_username():
             flash("Username changed successfully", "success")
             return redirect(url_for('user.profile'))
 
+        if request.method == 'GET':
+            return render_template('change_username.html')
     else:
         redirect(url_for('auth.login'))
 
 
-@user.route('/delete-user', methods=["GET", "POST"])
+@user.route('/delete-user', methods=["POST"])
 def delete_user():
     if 'email' in session:
         logged_user = db.query(User).filter_by(email=session['email']).first()
 
-        db.delete(logged_user)
-        db.commit()
+        if request.method == 'POST':
+            delete_password = request.form.get('password')
+            if logged_user.password == hashlib.sha256(delete_password.encode()).hexdigest():
+                db.delete(logged_user)
+                db.commit()
+                session.pop('email', None)
 
-        flash("User deleted successfully", "success")
-        return redirect(url_for('auth.login'))
+                flash("User deleted successfully", "success")
+                return redirect(url_for('auth.login'))
+            else:
+                flash("Incorrect password", "danger")
+                return redirect(url_for('user.profile'))
 
     else:
-        redirect(url_for('auth.login'))
+        return redirect(url_for('auth.login'))
